Add Inset and Spread to DropShadow struct, Blur to int64
Also: Use border for regular NativeView stroke
USe dropshadow and padding fo XImageView

+++ zfields/Field.go
@@ -591,7 +591,7 @@ func (f *Field) SetFromReflectValue(rval reflect.Value, sf reflect.StructField, 
 				if !got && f.Styling.DropShadow.Blur == 0 {
 					num, err := strconv.ParseFloat(part, 32)
 					if err != nil {
-						f.Styling.DropShadow.Blur = float32(num)
+						f.Styling.DropShadow.Blur = num
 						got = true
 					}
 				}
